Update field_sortable.php
Field_sortable updated to include label even when input mode is in use. I found a little bit confusing to not include the labels of the inputs, because when you have filled all the inputs, it could be hard to know what its the meaning of that input. 
For example: I have a sortable field for the social networks. The user must fill in the url of each social network he wants to include in the theme. Once he has done it, in some cases, an unexperienced user for example, the url could not give enough feedback to know which social network's is that url. Include a label is the proposed solution.